requirements.txt
astroid==2.3.3
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
idna==2.8
importlib-metadata==0.23
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.4.2
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pyparsing==2.4.4
python-dotenv==0.10.3
six==1.13.0
toml==0.10.0
tox==3.14.0
typed-ast==1.4.0
urllib3==1.25.7
virtualenv==16.7.7
wrapt==1.11.2
zipp==0.6.0
